+++ core/src/main/java/com/es/core/cart/CartService.java
@@ -6,9 +6,9 @@ public interface CartService {
 
     Cart getCart();
 
-    void addPhone(Long phoneId, Long quantity);
+    void addPhone(Long phoneId, Integer quantity);
 
-    void update(Map<Long, Long> items);
+    void update(Map<Long, Integer> items);
 
     void remove(Long phoneId);
     void clear();
